Refactored into multiple files
